add ctime

+++ clean.bat
@@ -1,1 +1,3 @@
-rmdir /s /q build
+@echo off
+IF EXIST build rmdir /s /q build
+echo ==== CLEAN ====
